Improve prompt selection
Do not select the prompt if the first candidate equals the input after the
completion boundary. This avoids jumping between candidate and prompt. A
candidate matching the input after the completion boundary will always be moved
first. This is a refinement of the previous commit, see #4. It is getting a bit
complicated unfortunately. Thank you for talking me into this, @oantolin...