Distribute the extensions together with the Vertico package (Fix #83)
This is the easiest approach in the long term. The extensions are still separate
from the main `vertico` feature and it will remain possible to manually install
`vertico.el` separately if only this file is desired. It turned out that
extensions are relatively robust, so there are essentially no downsides if
we distribute them together with the package.

(ELPA does not yet distribute the extensions with the vertio package as of 2021-11-30,
but the next release will include them.)