Document vertico-multiform
The `vertico-multiform` extension solves an issue I wanted to have fixed for
quite some time: It makes it possible to configure the Vertico display or
arbitrary modes per command or completion category, e.g., turning on the
`vertico-grid-mode` only for the command `find-file`. Until now it has only been
possible to turn on `vertico-grid-mode` or `vertico-buffer-mode` globally, which
wastes a lot of flexibility. The `vertico-multiform` package is still
experimental - turning modes on and off when entering and leaving recursive
minibuffers may introduce edge cases.

There is an interesting use case described by @abcdw described in
https://github.com/minad/consult/issues/486. Andrew uses @protesilaos' Mct and
configures it separately for each command. For commands like `consult-imenu`,
`consult-outline`, etc. he prefers a separate large buffer since the command is
meant to navigate within the current buffer. For buffer switching commands like
`find-file`, `switch-to-buffer` the selected window can be reused instead, keeping
the window layout intact.