Add vertico-activate-functions
A new variable called `vertico-activate-functions` is added. This variable is a list of
symbols that can be used to fine-tune which functions should actually activate the vertico
UI. It can also be a special value `(t)` that makes it activate always -- that is the
default behavior, which means it's "backward compatible".