vertico-insert: Always insert first candidate, if prompt is selected
I just tagged a stable release, so we can safely try a new behavior for a while.
The condition gets simpler, that's great. ;) Possible downside is that the
behavior may lead to accidental insertion of the first candidate even if the
candidate is not preferred.

By always inserting the first candidate one can repeatedly navigate directory
hierarchies as long as the completion candidates are sorted by
`vertico-sort-history-length-alpha`. `savehist-mode` remembers the history of
recent candidates across Emacs restarts. An alternative is the Prescient package
which sorts by frecency and which can also persist the selection history.

To give more context, there are multiple reasonable possibilities for
`vertico-insert` if the prompt line is selected:

- Use `completion-try-completion` to expand. This behavior is used by default
  completion or by my Corfu in-buffer completion package.
- Ignore TAB (old behavior, before #242)
- If there is only a single candidate, insert it (before this commit)
- If there is at least one candidate, insert first candidate (this commit)

See #272 and #237 for discussion. cc @raxod502